Move metrics tracker from shared utils to plugin reporting
Relocate createMetricsTracker from src/shared/utils/metrics-tracker.js to src/plugin/src/reporting/metrics-tracker.js because the tracker is only consumed by the plugin's identifier-case and project index modules.
