--- src/shared/metrics-tracker.js
@@ -1,1 +1,0 @@
-export { createMetricsTracker } from "./utils/metrics-tracker.js";
